Fixes Issue #12.
Restructure package to remove namespace confusion. All
randomized/conpressed algorithms have been moved into each of their
corresponding deterministic modules (ie rdmd is in the module dmd).
