+++ MANIFEST.in
@@ -1,5 +1,4 @@
 recursive-include data *
-recursive-include tests *
 recursive-include tutorial *
 recursive-include ristretto *.c *.pyx
