+++ ristretto/__init__.py
@@ -26,39 +26,41 @@ if __RISTRETTO_SETUP__:
     # yet be completed
     sys.stderr.write('Partial import of ristretto during the build process.\n')
 else:
-    # Matrix Factorization
-    from ristretto.mf import rsvd
-    from ristretto.mf import rsvd_single
-    from ristretto.mf import csvd
-
-    from ristretto.mf import cur
-    from ristretto.mf import rcur
-
-    from ristretto.mf import interp_decomp
-    from ristretto.mf import rinterp_decomp
-    from ristretto.mf import rinterp_decomp_qb
-
-    from ristretto.mf import rlu
-
-    from ristretto.mf import rqb
-
-    from ristretto.mf import reigh
-    from ristretto.mf import reigh_nystroem
-    from ristretto.mf import reigh_nystroem_col
-
-    # Dynamic Mode Decomposition
-    from ristretto.dmd import dmd
-    from ristretto.dmd import rdmd
-    from ristretto.dmd import rdmd_single
-
-    #Nonnegative MF
-    from ristretto.nmf import nmf
-    from ristretto.nmf import rnmf
-
-    #PCA
-    from ristretto.pca import robspca
-    from ristretto.pca import rspca
-    from ristretto.pca import spca
-
-    #Utilities
-    from ristretto import utils
+    __all__ = ['cur', 'dmd', 'eigen', 'interp_decomp', 'lu', 'nmf', 'pca',
+               'qb', 'sketch', 'svd', 'utils']
+#    # Matrix Factorization
+#    from ristretto.mf import rsvd
+#    from ristretto.mf import rsvd_single
+#    from ristretto.mf import csvd
+#
+#    from ristretto.mf import cur
+#    from ristretto.mf import rcur
+#
+#    from ristretto.mf import interp_decomp
+#    from ristretto.mf import rinterp_decomp
+#    from ristretto.mf import rinterp_decomp_qb
+#
+#    from ristretto.mf import rlu
+#
+#    from ristretto.mf import rqb
+#
+#    from ristretto.mf import reigh
+#    from ristretto.mf import reigh_nystroem
+#    from ristretto.mf import reigh_nystroem_col
+#
+#    # Dynamic Mode Decomposition
+#    from ristretto.dmd import dmd
+#    from ristretto.dmd import rdmd
+#    from ristretto.dmd import rdmd_single
+#
+#    #Nonnegative MF
+#    from ristretto.nmf import nmf
+#    from ristretto.nmf import rnmf
+#
+#    #PCA
+#    from ristretto.pca import robspca
+#    from ristretto.pca import rspca
+#    from ristretto.pca import spca
+#
+#    #Utilities
+#    from ristretto import utils
