+++ ristretto/eigen.py
@@ -10,8 +10,8 @@ from __future__ import division, print_function
 import numpy as np
 from scipy import linalg
 
-from ..sketch import sketch
-from ..utils import conjugate_transpose
+from .sketch import sketch
+from .utils import conjugate_transpose
 
 _VALID_DTYPES = (np.float32, np.float64, np.complex64, np.complex128)
