+++ ristretto/tests/test_cur.py
@@ -1,12 +1,15 @@
 import numpy as np
 
-from ristretto.mf import cur, rcur
-
+from ristretto.cur import cur
+from ristretto.cur import rcur
 
 atol_float32 = 1e-4
 atol_float64 = 1e-8
 
 
+# =============================================================================
+# cur function
+# =============================================================================
 def test_cur():
     m, k = 100, 10
     A = np.array(np.random.randn(m, k), np.float64)
@@ -24,6 +27,9 @@ def test_cur():
     assert relative_error < 1e-4
 
 
+# =============================================================================
+# rcur function
+# =============================================================================
 def test_rcur():
     m, k = 100, 10
     A = np.array(np.random.randn(m, k), np.float64)
