+++ ristretto/tests/test_eigen.py
@@ -1,12 +1,16 @@
 import numpy as np
 
-from ristretto.mf import reigh, reigh_nystroem, reigh_nystroem_col
-
+from ristretto.eigen import reigh
+from ristretto.eigen import reigh_nystroem
+from ristretto.eigen import reigh_nystroem_col
 
 atol_float32 = 1e-4
 atol_float64 = 1e-8
 
 
+# =============================================================================
+# reigh function
+# =============================================================================
 def test_reigh_float64():
     m, k = 100, 10
     A = np.array(np.random.randn(m, k), np.float64)
@@ -32,7 +36,9 @@ def test_reigh_complex128():
     assert percent_error < atol_float64
 
 
-
+# =============================================================================
+# reig_nystroem function
+# =============================================================================
 def test_reig_nystroem_float64():
     m, k = 20, 10
     A = np.array(np.random.randn(m, k), np.float64)
@@ -57,6 +63,9 @@ def test_reig_nystroem_complex128():
     assert percent_error < atol_float64
 
 
+# =============================================================================
+# reig_nystroem_col function
+# =============================================================================
 def test_reig_nystroem_col_float64():
     m, k = 20, 10
     A = np.array(np.random.randn(m, k), np.float64)
