+++ ristretto/tests/test_qb.py
@@ -1,12 +1,14 @@
 import numpy as np
 
-from ristretto.mf import rqb
-
+from ristretto.qb import rqb
 
 atol_float32 = 1e-4
 atol_float64 = 1e-8
 
 
+# =============================================================================
+# rqb function
+# =============================================================================
 def test_rqb_float64():
     m, k = 100, 10
     A = np.array(np.random.randn(m, k), np.float64)
